chore: Rename files to match their exports
