+++ web/src/routes/_rallyists/_withUserProviderNoAutoAnonymous/settings.tsx
@@ -7,8 +7,8 @@ import enFlag from "@/assets/languages/en.svg";
 import frFlag from "@/assets/languages/fr.svg";
 import { ButtonLink } from "@/components/ButtonLink.tsx";
 import { Checkbox } from "@/components/Checkbox.tsx";
+import { CreateAccountForm } from "@/components/CreateAccountForm.tsx";
 import { Header } from "@/components/Header.tsx";
-import { CreateAccountForm } from "@/components/createAccountForm.tsx";
 import { getPrefs, setPref } from "@/lib/appwrite.ts";
 import { useLogout } from "@/lib/hooks/useLogout.ts";
 import { useUser } from "@/lib/hooks/useUser.ts";
